fix(plugin): Player wasn't killed when logging out
When running the `logout` command, the player should be put into
derelict -- this is to prevent the player from spawning, sending chat
messages, and from placing blocks. However, the line of code to kill the
player wasn't present, so the player could switch to the derelict team
while also keeping their unit.

Not a huge issue tbh, but I'd prefer to deter players from finding ways
to make moderation more annoying.