refactor(plugin): Updated commands structure and streamlined configuration reloading
Updated the structure of the `commands` package. Additionally, the
process of applying updates to features that are dependent on the
configuration has been streamlined. So, instead of reloading the
database and reloading configuration dependent features via separate
methods, you can now reload the database and configuration dependent
features via a singular method.

BREAKING CHANGE: Updated package structure for commands